Separate data generation and distributed training
